+++ packages/featureserver/CHANGELOG.md
@@ -2,14 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [2.11.1] - 04-18-2018
 ### Fixed
 * default to templated value of display field if not provided in metadata and an `empty` fields array
 
 ### Added
 * if request includes `returnGeometry=false` and `outSR=<EPSG>`, delete the `outSR` param since no geometry will be present to reproject
 
-## [2.11.0] - 04-16-2018
+## [2.11.0] - 04-17-2018
 ### Added
 * provider attributes referenced by metadata `idField` are maintained as separate field in addition to OBJECTID when creating ESRI json
 * warnings when a provider's `idField` is not set or references non- or out-of-range integer values
@@ -279,6 +279,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2017-01-25
 * Initial Release
 
+[2.11.1]: https://github.com/koopjs/featureserver/compare/v2.11.0...v2.11.1
 [2.11.0]: https://github.com/koopjs/featureserver/compare/v2.10.2...v2.11.0
 [2.10.2]: https://github.com/koopjs/featureserver/compare/v2.10.1...v2.10.2
 [2.10.1]: https://github.com/koopjs/featureserver/compare/v2.10.0...v2.10.1